requirements.txt
flask
pytesseract
opencv-python
pillow
pdf2image
pandas
openpyxl
gunicorn